pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "errata"
version = "0.1.0"
description = "A/B testing tool for agentic AI models"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "openai>=1.65.0",
    "google-generativeai>=0.8.0",
    "rich>=13.9.0",
    "prompt-toolkit>=3.0.48",
    "pydantic-settings>=2.7.0",
    "pyperclip>=1.9.0",
    "toml>=0.10.2",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-mock>=3.14.0",
]

[project.scripts]
errata = "errata.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/errata"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
